+++ server/out/providers/idl-document-symbol-manager.js
@@ -137,18 +137,21 @@ class IDLDocumentSymbolManager {
             }
             else {
                 // init index
-                let idx = -1;
+                let remove = [];
                 // TODO: for more than function names, use some rigor for matching symbols
                 // as we can have multiple symbols with the same name that are different entities
                 // compare each document with a match for that symbol
                 this.symbols[key].forEach((docMatch, dIdx) => {
                     if (docMatch.uri === uri) {
-                        idx = dIdx;
+                        remove.push(dIdx);
                     }
                 });
                 // delete if we found a match
-                if (idx !== -1) {
-                    this.symbols[key].splice(idx, 1);
+                if (remove.length > 0) {
+                    remove.reverse();
+                    remove.forEach(idx => {
+                        this.symbols[key].splice(idx, 1);
+                    });
                 }
             }
         });
@@ -158,6 +161,7 @@ class IDLDocumentSymbolManager {
         return __awaiter(this, void 0, void 0, function* () {
             // check if we need to remove it from the symbol list
             if (this.moizes.documentSymbols.has([uri])) {
+                this.connection.console.log(JSON.stringify("Have cached symbols"));
                 // get our symbols and clean up
                 const symbols = yield this.get.documentSymbols(uri);
                 // clean up our symbol information