+++ server/src/core/problems.interface.ts
@@ -1,0 +1,5 @@
+import { Diagnostic } from "vscode-languageserver";
+
+export interface IProblems {
+  [key: string]: Diagnostic[];
+}